python_version/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
# For Roblox API interactions
roblox.py>=0.4.2
requests>=2.26.0